commit branch feature/all-post

+++ app/Http/Controllers/Admin/PostController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ApprovalPendingPostRequest;
 use App\Http\Requests\StorePostRequest;
+use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
 use App\Services\CategoryService;
 use App\Services\PostService;
@@ -24,6 +26,8 @@ class PostController extends Controller
 
     public function index(): Response
     {
+        $this->authorize('view', Post::class);
+
         return response()->view('admin.post.index', [
             'posts' => $this->postService->getPostsPagination(10),
             'title' => 'Haalaman Admin Post'
@@ -43,10 +47,13 @@ class PostController extends Controller
     public function store(StorePostRequest $request): RedirectResponse
     {
         try {
+            $this->authorize('create', Post::class);
+
             Log::info('Post created successfully!', [
                 'user_id' => Auth::user()->id,
                 'payload' => $request->except('body')
             ]);
+
             $this->postService->createPost($request->validated());
             return to_route('admin.posts.index')->with('success', 'Post created successfully!');
         } catch (Exception $e) {
@@ -57,33 +64,39 @@ class PostController extends Controller
 
     public function show(Post $post): Response
     {
+        $this->authorize('view', $post);
+
         return response()->view('admin.post.show', [
-            'post' => $post->load(['author', 'category'])
+            'post' => $this->postService->findBySlug($post)
         ]);
     }
 
 
     public function edit(Post $post): Response
     {
+        $this->authorize('view', $post);
+
         return response()->view('admin.post.edit', [
-            'post' => $post->load('category'),
+            'post' => $this->postService->findBySlug($post),
             'categories' => $this->categoryService->getCategories(),
         ]);
     }
 
 
-    public function update(StorePostRequest $request, Post $post): RedirectResponse
+    public function update(ApprovalPendingPostRequest $request, Post $post): RedirectResponse
     {
         try {
-            Log::info('Post update successfully!', [
+            $this->authorize('update', $post);
+
+            Log::info('Post approved successfully!', [
                 'user_id' => Auth::user()->id,
                 'payload' => $request->except('body')
             ]);
-            $this->postService->updatePost($post, $request->validated());
-            return to_route('admin.posts.index')->with('success', 'Post updated successfully!');
+            $this->postService->approvalPendingPost($post, $request->validated());
+            return redirect()->back()->with('success', 'Post approved successfully!');
         } catch (Exception $e) {
-            Log::error('Post update failed! : ' . $e->getMessage());
-            return redirect()->back()->withInput()->with('error', 'Post update failed!');
+            Log::error('Post approve failed! : ' . $e->getMessage());
+            return redirect()->back()->withInput()->with('error', 'Post approve failed!');
         }
     }
 
@@ -91,10 +104,13 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         try {
+            $this->authorize('forceDelete', $post);
+
             Log::info('Post deleted successfully!', [
                 'user_id' => Auth::user()->id,
                 'item_deleted' => $post->id
             ]);
+
             $this->postService->removePost($post);
             return to_route('admin.posts.index')->with('success', 'Post deleted successfully!');
         } catch (Exception $e) {
